Remove mount versions from `version` command
A `version` command is generally expected to be just a simple print of
the application version.  In particular, it should not trigger the
auto-sync that occurs when loading the `gopass` configuration in order
to show the command versions for the configured mounts.  Instead, the
version info for the mount points is now displayed by a separate
`mounts versions` sub-command.
