+++ internal/action/commands.go
@@ -680,6 +680,16 @@ Use the "git init" command if the store does not yet have a git repository.`,
 					Action:       s.MountRemove,
 					BashComplete: s.MountsComplete,
 				},
+				{
+					Name:    "versions",
+					Aliases: []string{"version"},
+					Usage:   "Display mount provider versions",
+					Description: "" +
+						"This command displays version information of important external " +
+						"commands used by the configured password store mounts.",
+					Before: s.IsInitialized,
+					Action: s.MountsVersions,
+				},
 			},
 		},
 		{
@@ -920,9 +930,7 @@ Use the "git init" command if the store does not yet have a git repository.`,
 			Name:  "version",
 			Usage: "Display version",
 			Description: "" +
-				"This command displays version and build time information " +
-				"along with version information of important external commands. " +
-				"Please provide the output when reporting issues.",
+				"This command displays version and build time information.",
 			Action: s.Version,
 		},
 	}
